Level 1 Segment 3 Polishing
Changed terrain material and layout
Changed fenced areas and shapes.
Added dinosaur assets.
Created new placeholders for the trees.
Added the car park to map out where it would have to sit.
