+++ Assets/NewLayer.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1670a833bde45514ca6b09f4105737de
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8574412962073106934
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
