+++ Assets/_Developers/Art/Prop/Dinosaurs/Parasaur/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b38f16c55dbbbe1499d3cc0ebc430b92
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
