+++ Assets/_Developers/Art/Prop/Dinosaurs/Parasaur/Materials/Parasaur_AlbedoTransparency.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d4acbb18cd0f5a547bdf15c1a32df0be
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
