+++ Assets/_Developers/Art/Prop/Dinosaurs/Pterodactyl/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 394873f812fea604e968b5bcb40887fe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
