+++ Assets/_Developers/Art/Prop/Dinosaurs/Pterodactyl/Materials/Pterodactyl_AlbedoTransparency.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9a9a89fef09cc1b40a9baef1211047a8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
