+++ Assets/_Developers/Art/Prop/Dinosaurs/Spinosaurus/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3439fd876b657c047888cf9dbcbbe545
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
