+++ Assets/_Developers/Art/Prop/Dinosaurs/Spinosaurus/Materials/Spinosaurus_AlbedoTransparency.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 04560b1458bd71e4f903c3467e14195e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
