+++ Assets/_Developers/Art/Prop/Dinosaurs/Stegosaurus/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fbbe76b58e204d0418b5d36607923262
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
