+++ Assets/_Developers/Art/Prop/Dinosaurs/Stegosaurus/Materials/Stegosaurus_AlbedoTransparency.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dd202cf697e20ab48a4307755ec17db6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
