--- Assets/_Developers/LD/Level 1/CharlieY.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7b7024cb64a0150459ca3424cd417bee
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
