--- Assets/_Developers/LD/Level 1/CharlieY/CYLevelStreaming.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aaef4310ab17c674f93fcd982352acd8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
