--- Assets/_Developers/LD/Level 1/CharlieY/PvELevel1Part3.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 37e0b6169241c1e4088e8703722afdab
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
