+++ Assets/_Developers/LD/Level 1/EmilyH.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b7024cb64a0150459ca3424cd417bee
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
