+++ Assets/_Developers/LD/Level 1/EmilyH/CYLevelStreaming.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aaef4310ab17c674f93fcd982352acd8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
