+++ Assets/_Developers/LD/Level 1/EmilyH/PlaceHolderStuff.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b94ed59d81764354fb135e8f479f537f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
