+++ Assets/_Developers/LD/Level 1/EmilyH/PlaceHolderStuff/PlaceHOlderFence.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b68ae6c15d34b7f408ba88292e4be09e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
