+++ Assets/_Developers/LD/Level 1/EmilyH/PvELevel1Part3.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 37e0b6169241c1e4088e8703722afdab
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
